+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.19.0]
+
 * doc: add documentation on the remote backend (#90)
 * feat: add seed to mnist example (#103)
 * Drop Python 3.7 support (#96)